refactor: features 하위 컴포넌트들을 도메인 기반 폴더로 이동
